+++ lib/ood/user_map.lua
@@ -4,17 +4,14 @@
   Given a request and authenticated user, map this user to a system-level user
   using the supplied shell command.
 --]]
-function map(r, user_map_cmd)
-  -- read in authenticated user
-  local auth_user = r:escape(r.user)
-
+function map(r, user_map_cmd, remote_user)
   -- run user_map_cmd and read in stdout
   local now = r:clock()
-  local handle = io.popen(user_map_cmd .. " '" .. auth_user .. "'")
+  local handle = io.popen(user_map_cmd .. " '" .. r:escape(remote_user) .. "'")
   sys_user = handle:read()
   handle:close()
   time_user_map = (r:clock() - now)/1000.0
-  r:info("Mapped '" .. r.user .. "' => '" .. (sys_user or "") .. "' [" .. time_user_map .. " ms]")
+  r:info("Mapped '" .. remote_user .. "' => '" .. (sys_user or "") .. "' [" .. time_user_map .. " ms]")
 
   -- failed to map if returns empty string
   if not sys_user or sys_user == "" then